# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: RFP Agent API
  version: 1.0.0
  description: |
    **RFP Agent** is an AI-powered platform for automated RFP discovery, proposal generation, and submission management.

    ## Features
    - 🔍 **Automated RFP Discovery**: Scan government portals for new opportunities
    - 🤖 **AI-Powered Proposal Generation**: Generate compliant proposals using GPT-5
    - 📊 **3-Tier Agent System**: Orchestrator, Managers, and Specialist agents
    - 📄 **Document Intelligence**: Parse and analyze RFP documents
    - 🚀 **Automated Submission**: Submit proposals through portal automation
    - 📈 **Analytics & Monitoring**: Track performance and success rates

    ## Architecture
    The platform uses a sophisticated 3-tier agentic system:
    - **Tier 1**: Primary Orchestrator (1 agent)
    - **Tier 2**: Managers - Portal, Proposal, Research (3 agents)
    - **Tier 3**: Specialists - Scanner, Monitor, Generator, Checker, Processor, Market Analyst, Historical Analyzer (7 agents)

    ## Authentication
    Currently, the API uses session-based authentication. API key support is planned for future releases.

  contact:
    name: RFP Agent Support
    email: support@rfpagent.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.rfpagent.com/api
    description: Production server
...
paths:
...
  /proposals/rfp/{rfpId}:
    get:
      tags: [Proposals]
      summary: Get proposals for RFP
      description: Retrieve all proposals associated with a specific RFP
      operationId: getProposalsForRFP
      parameters:
        - name: rfpId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Proposal'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /proposals/{id}:
    put:
      tags: [Proposals]
      summary: Update proposal
      description: Update proposal details
      operationId: updateProposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProposalRequest'
      responses:
        '200':
          description: Proposal updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proposal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Proposals]
      summary: Delete proposal
      description: Remove a proposal from the system
      operationId: deleteProposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Proposal deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /proposals/enhanced/generate:
    post:
      tags: [Proposals]
      summary: Generate enhanced proposal
      description: Start AI-powered proposal generation with advanced features
      operationId: generateEnhancedProposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - rfpId
              properties:
                rfpId:
                  type: string
                  format: uuid
                companyProfileId:
                  type: string
                  format: uuid
                options:
                  type: object
                  properties:
                    generatePricing:
                      type: boolean
                      default: true
                    generateCompliance:
                      type: boolean
                      default: true
                    proposalType:
                      type: string
                      enum: [standard, technical, cost]
                    qualityThreshold:
                      type: number
                      minimum: 0
                      maximum: 1
                      default: 0.8
            example:
              rfpId: "550e8400-e29b-41d4-a716-446655440000"
              companyProfileId: "660e8400-e29b-41d4-a716-446655440000"
              options:
                generatePricing: true
                generateCompliance: true
                proposalType: "technical"
                qualityThreshold: 0.85
      responses:
        '200':
          description: Proposal generation started
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  sessionId:
                    type: string
                  message:
                    type: string
              example:
                success: true
                sessionId: "enhanced_550e8400_1234567890"
                message: "Enhanced proposal generation started"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /proposals/pipeline/generate:
    post:
      tags: [Proposals]
      summary: Generate proposals in pipeline
      description: Generate multiple proposals in parallel or sequential pipeline
      operationId: generateProposalPipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - rfpIds
                - companyProfileId
              properties:
                rfpIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                companyProfileId:
                  type: string
                  format: uuid
                priority:
                  type: integer
                  minimum: 1
                  maximum: 10
                  default: 5
                parallelExecution:
                  type: boolean
                  default: true
                options:
                  type: object
      responses:
        '200':
          description: Pipeline started successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  pipelines:
                    type: array
                    items:
                      type: object
                  rfpCount:
                    type: integer
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /proposals/{id}/approve:
    post:
      tags: [Proposals]
      summary: Approve proposal
      description: Mark a proposal as approved for submission
      operationId: approveProposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                approvedBy:
                  type: string
                notes:
                  type: string
      responses:
        '200':
          description: Proposal approved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  proposal:
                    $ref: '#/components/schemas/Proposal'
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # PORTAL ENDPOINTS
  # ============================================================================
...
components:
  schemas:
...
    Proposal:
      type: object
      required:
        - id
        - rfpId
        - status
      properties:
        id:
          type: string
          format: uuid
        rfpId:
          type: string
          format: uuid
        content:
          type: object
        narratives:
          type: object
        pricingTables:
          type: object
        forms:
          type: object
        attachments:
          type: object
        proposalData:
          type: object
        estimatedCost:
          type: string
        estimatedMargin:
          type: string
        status:
          type: string
          enum: [draft, review, approved, submitted]
        generatedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateProposalRequest:
      type: object
      properties:
        content:
          type: object
        narratives:
          type: object
        pricingTables:
          type: object
        status:
          type: string
          enum: [draft, review, approved, submitted]
        estimatedCost:
          type: string
        estimatedMargin:
          type: string
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: object
          example:
            error: "Invalid input data"
            details:
              - field: "url"
                message: "Must be a valid URL"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: "Resource not found"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: "Internal server error"

  # ============================================================================
  # SECURITY SCHEMES
  # ============================================================================
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session-based authentication using cookies

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication (planned for future release)

# Global security - currently using session-based auth
security:
  - sessionAuth: []